move spamFilter interface
