+++ packages/lib/src/spam-filter/filter/EthBalanceFilter.ts
@@ -1,6 +1,6 @@
 import { ethers } from 'ethers';
-import { SpamFilter } from '..';
 import { DeliveryInformation } from '../../messaging';
+import { SpamFilter } from './SpamFilter';
 
 export interface EthBalanceFilterSettings {
     ethHigherOrEqualThan: string;
