+++ packages/lib/src/spam-filter/filter/SpamFilter.ts
@@ -1,0 +1,5 @@
+import { DeliveryInformation } from '../../messaging';
+
+export interface SpamFilter {
+    filter(e: DeliveryInformation): Promise<boolean>;
+}
